refactor(workers): 二十四節気メタを共通モジュールに切り出し
F-1.3b 着手の前段。Generator と Judge の両方で SOLAR_TERM_META を
使うため、solar_term_meta.js を新設して DRY 化。
anthropic.js の SOLAR_TERM_MAP は META に吸収して削除。
既存テスト 97 件 全 pass。
